More unit tests and clean up of ODentry code, ODentry no longer has a subindex field, although one still exists its now readonly and is an accessor that gets parent and looks up in Dictionary. This means that the Dictionary key/value is authoritative in the subindex position/ordering. Any code that set subindex has been removed, any code that for looped on sub index has been chanegd to foreach on keyvaluepair of subobjects Some places in GUI really need to know a subobjects index from subobject alone so accessor was added

+++ Tests/EDSParserTests.cs
@@ -99,7 +99,7 @@ CompactSubObj=9
             if (sub.PDOtype != PDOMappingType.no)
                 throw (new Exception("parseEDSentry() CompactSubObj incorrect generation"));
 
-            for (int x = 1; x < 10; x++)
+            for (UInt16 x = 1; x < 10; x++)
             {
                 if (!od.containssubindex(x))
                     throw (new Exception("parseEDSentry() CompactSubObj incorrect generation"));
